+++ project/Settings.scala
@@ -67,7 +67,19 @@ object Settings {
   })
 
   lazy val commonDependencyOverrides = Seq(
-    "org.apache.commons" % "commons-compress" % "1.26.0"
+    "org.apache.commons" % "commons-compress" % "1.26.0",
+    // Needed to override org.broadinstitute.dsde:rawls-model_2.13
+    "io.netty" % "netty-handler" % "4.1.110.Final",
+    "io.netty" % "netty-codec-http" % "4.1.110.Final",
+    "org.eclipse.jetty" % "jetty-http" % "12.0.12",
+    "commons-net" % "commons-net" % "3.9.0",
+    "org.apache.commons" % "commons-text" % "1.10.0",
+    "xalan" % "xalan" % "2.7.3",
+    "org.apache.httpcomponents" % "httpclient" % "4.5.14",
+    // Needed to override opencensus
+    "io.grpc" % "grpc-protobuf" % "1.59.0",
+    // Needed to override kubernetes client
+    "com.google.protobuf" % "protobuf-java" % "3.25.5"
   )
 
   val scala213 = "2.13.12"
@@ -111,14 +123,14 @@ object Settings {
   val googleSettings = commonSettings ++ List(
     name := "workbench-google",
     libraryDependencies ++= googleDependencies,
-    version := createVersion("0.33"),
+    version := createVersion("0.34"),
     coverageExcludedPackages := ".*HttpGoogle.*DAO.*"
   ) ++ publishSettings
 
   val google2Settings = commonSettings ++ List(
     name := "workbench-google2",
     libraryDependencies ++= google2Dependencies,
-    version := createVersion("0.36")
+    version := createVersion("0.37")
   ) ++ publishSettings
 
   val azureSettings = commonSettings ++ List(
@@ -130,19 +142,19 @@ object Settings {
   val openTelemetrySettings = commonSettings ++ List(
     name := "workbench-openTelemetry",
     libraryDependencies ++= openTelemetryDependencies,
-    version := createVersion("0.8")
+    version := createVersion("0.9")
   ) ++ publishSettings
 
   val errorReportingSettings = commonSettings ++ List(
     name := "workbench-error-reporting",
     libraryDependencies ++= errorReportingDependencies,
-    version := createVersion("0.8")
+    version := createVersion("0.9")
   ) ++ publishSettings
 
   val serviceTestSettings = commonSettings ++ List(
     name := "workbench-service-test",
     libraryDependencies ++= serviceTestDependencies,
-    version := createVersion("5.0")
+    version := createVersion("5.1")
   ) ++ publishSettings
 
   val notificationsSettings = commonSettings ++ List(
@@ -154,7 +166,7 @@ object Settings {
   val oauth2Settings = commonSettings ++ List(
     name := "workbench-oauth2",
     libraryDependencies ++= oauth2Dependencies,
-    version := createVersion("0.8")
+    version := createVersion("0.9")
   ) ++ publishSettings
 
   val rootSettings = commonSettings ++ noPublishSettings ++ noTestSettings